Add utility functions and update README; include new images for analysis
